Alpha implementation of max peak kalmans in Rho
This has eliminated Sigma and Kalman Matrix structure
Updates probabilities relative to max peak kalmans so state are still used

+++ Programs/Algorithms/Image Processing/tau/tau+/a3/tau+/global_types.h
@@ -16,7 +16,7 @@ extern "C" {
 #include "test_setup.h"
 #include <stdint.h>
 
-static void cma( double new_val, double *avg, int num ) { *avg += ( new_val - *avg ) / ( num + 1 ); }
+static void cma( double new_val, double *avg, int num ) { *avg += ( new_val - *avg ) / (double)( num + 1 ); }
 
 #ifdef __cplusplus
 }
